Move the module into users.py and make required changes
