+++ tg_bot/modules/users.py
@@ -1,3 +1,5 @@
+from io import BytesIO
+
 from time import sleep
 from typing import Optional
 
@@ -9,6 +11,7 @@ from telegram.ext.dispatcher import run_async
 
 import tg_bot.modules.sql.users_sql as sql
 from tg_bot import dispatcher, OWNER_ID, LOGGER
+from tg_bot.modules.helper_funcs.filters import CustomFilters
 
 USERS_GROUP = 4
 
@@ -84,6 +87,19 @@ def log_user(bot: Bot, update: Update):
                         msg.forward_from.username)
 
 
+@run_async
+def chats(bot: Bot, update: Update):
+    chats = sql.get_all_chats() or []
+    chatfile = 'List of chats.\n'
+    for chat in chats:
+        chatfile += "{} - {}\n".format(chat.chat_name, chat.chat_id)
+
+    with BytesIO(str.encode(chatfile)) as output:
+        output.name = "chatlist.txt"
+	update.effective_message.reply_document(document=output, filename="chatlist.txt",
+                                                caption="Here is the list of chats in my database.")
+
+
 def __user_info__(user_id):
     if user_id == dispatcher.bot.id:
         return """I've seen them in... Wow. Are they stalking me? They're in all the same places I am... oh. It's me."""
@@ -106,6 +122,8 @@ __mod_name__ = "Users"
 
 BROADCAST_HANDLER = CommandHandler("broadcast", broadcast, filters=Filters.user(OWNER_ID))
 USER_HANDLER = MessageHandler(Filters.all & Filters.group, log_user)
+CHATLIST_HANDLER = CommandHandler("chats", chats, filters=CustomFilters.sudo_filter)
 
 dispatcher.add_handler(USER_HANDLER, USERS_GROUP)
 dispatcher.add_handler(BROADCAST_HANDLER)
+dispatcher.add_handler(CHATLIST_HANDLER)
